Fixed a bug in parsing of arcs in lib files.
Some parts in the transformer library including the P0926NL cause
the parse_arc function to segfault. This is because there is an
assert check on the start and end locations of the arc. Removing
this assert stops the fault and the arcs are still rendered,
although the are mirrored relative to their intended orientation.

My intended use case does not use the transformer library so I
decided to just ignore this issue.